+++ libzappy_net/src/zappy_net_ringbuf_utils.c
@@ -6,7 +6,6 @@
 */
 
 #include "../include/zappy_net_ringbuf.h"
-#include <stdint.h>
 
 int zn_count_newlines(const uint8_t *data, size_t len)
 {
@@ -40,21 +39,4 @@ size_t zn_ringbuf_read_available(const zn_ringbuf_t *rb)
         return rb->write_pos - rb->read_pos;
     else
         return rb->capacity - rb->read_pos + rb->write_pos;
-}
-
-int zn_ringbuf_is_full(const zn_ringbuf_t *rb)
-{
-    return zn_ringbuf_write_available(rb) == 0;
-}
-
-int zn_ringbuf_is_empty(const zn_ringbuf_t *rb)
-{
-    return (rb == NULL || rb->buffer == NULL || rb->is_empty);
-}
-
-size_t zn_ringbuf_count_lines(const zn_ringbuf_t *rb)
-{
-    if (rb == NULL)
-        return 0;
-    return rb->line_count;
 }
